+++ Sandswept/Enemies/GammaConstruct/Skills/DeathState.cs
@@ -1,0 +1,34 @@
+using System;
+using R2API.Utils;
+
+namespace Sandswept.Enemies.GammaConstruct
+{
+    public class DeathState : GenericCharacterDeath
+    {
+        public static GameObject deathEffect = SpawnAndDeath.gammaConstructSpawnAndDeathVFX;
+
+        public override void CreateDeathEffects()
+        {
+            base.CreateDeathEffects();
+
+            characterBody.modelLocator.autoUpdateModelTransform = false;
+            cachedModelTransform.parent = null;
+
+            var boxes = cachedModelTransform.GetComponentsInChildren<HurtBox>(true);
+            for (int i = 0; i < boxes.Length; i++)
+            {
+                boxes[i].enabled = false;
+                boxes[i].gameObject.SetActive(false);
+            }
+            cachedModelTransform.GetComponent<RagdollController>().BeginRagdoll(Vector3.one * -5f);
+
+            EffectManager.SpawnEffect(deathEffect, new EffectData
+            {
+                origin = base.characterBody.corePosition,
+                scale = base.characterBody.bestFitRadius * 2f,
+            }, false);
+
+            Util.PlaySound("Play_item_lunar_specialReplace_explode", base.gameObject);
+        }
+    }
+}